// frontend/js/trueFalse-alkis/core-transfer/types/models.ts
import type { QuizCategory, StatementType } from "./enums"

interface TimeStamped {
	created: string
	modified: string
}

interface Activatable {
	is_active: boolean
}

export interface ExamSession extends TimeStamped {
	id: number
	year: number
	month: number
	month_display: string
}

export interface QuizAsset extends TimeStamped {
	id: number
	title: string
	image: string
}

interface QuizChoice {
	text?: string
	asset_id?: number
	is_correct: boolean
}

// αλλαγή απο core σε export
export interface TrueFalseContent {
	prompt_text?: string
	prompt_asset_id?: number
	choices: QuizChoice[]
}

interface MultipleChoiceContent {
	prompt_text?: string
	prompt_asset_id?: number
	choices: QuizChoice[]
}

export interface Statement extends TimeStamped, Activatable {
	id: number
	type: StatementType
	category: QuizCategory
	content: TrueFalseContent | MultipleChoiceContent
	exam_sessions: ExamSession[]
	exam_sessions_preview: string
}

interface DragAndDropColumn {
	title: string
	values: string[]
}

export type DragAndDropContent = [DragAndDropColumn, DragAndDropColumn]

export interface DragAndDrop extends TimeStamped, Activatable {
	id: number
	category: QuizCategory
	content: DragAndDropContent
	exam_sessions: ExamSession[]
	exam_sessions_preview: string
}

interface MatchingItem {
	id?: number
	matched_id?: number
	text: string
}

interface MatchingColumn {
	title: string
	items: MatchingItem[]
}

export type MatchingContent = [MatchingColumn, MatchingColumn]

export interface Matching extends TimeStamped, Activatable {
	id: number
	category: QuizCategory
	content: MatchingContent
	exam_sessions: ExamSession[]
	exam_sessions_preview: string
}

interface FillBlankChoice {
	text: string
	is_correct: boolean
}

interface FillBlankTextPart {
	text: string
	is_blank: boolean
	choices?: FillBlankChoice[]
}

interface FillBlankText {
	parts: FillBlankTextPart[]
}

interface FillInTheBlankContent {
	prompt_asset_id?: number
	show_answers_as_choices: boolean
	has_multiple_choices: boolean
	texts: FillBlankText[]
}

export interface FillInTheBlank extends TimeStamped, Activatable {
	id: number
	category: QuizCategory
	content: FillInTheBlankContent
	exam_sessions: ExamSession[]
	exam_sessions_preview: string
}

export interface Exam {
	true_false: [Statement]
	multiple_choice: [Statement]
	fill_in_the_blank: [FillInTheBlank]
	drag_and_drop: [DragAndDrop]
	matching: [Matching]
}

export type TrainingData = {
	id: number
	category: QuizCategory
	content:
		| FillInTheBlankContent
		| TrueFalseContent
		| MultipleChoiceContent
		| MatchingContent
		| DragAndDropContent
	quiz_type: string
}[]